refactor(workspace): split attach-indexed-db into three files
attach-indexed-db.ts grew three unrelated exports sharing only a return
type: the plain y-indexeddb wrapper, the custom encrypted protocol, and
the owner-scoped bulk delete. The SKILL doc references attach-indexed-db
as "the canonical 40-line example," but the 155-line encrypted variant
drowned that out.

Split into:
- attach-indexed-db.ts (~46 lines): attachIndexedDb + IndexedDbAttachment
- attach-encrypted-indexed-db.ts (~178 lines): attachEncryptedIndexedDb +
  EncryptedIndexedDbOptions
- wipe-owner-local-yjs-data.ts (~41 lines): wipeOwnerLocalYjsData + its
  local IDB factory types

attach-encrypted-indexed-db.ts imports IndexedDbAttachment from
attach-indexed-db.ts; no barrel re-exports. Direct importer paths
updated in attach-encryption.ts, the test file, and the barrel.
